refactor: Remove geometry flat file wrappers, complete subdirectory migration (Issue #320 Phase 2)
Completes the geometry module refactoring by removing flat file wrappers
and updating all imports to use subdirectory paths.

Changes:
- Remove wrapper files: simple_grid_1d.py, simple_grid.py, tensor_product_grid.py
- Update geometry/__init__.py to import from grids subdirectory
- Update all internal imports across codebase:
  * mfg_pde/ modules (4 files)
  * tests/ (7 files)
  * examples/ (2 files)
  * docs/ (2 files)
- Update relative imports in geometry/ module itself (3 files)
- All 305 geometry tests passing

Migration summary:
- simple_grid_1d.py → grids/grid_1d.py (canonical)
- simple_grid.py → grids/grid_2d.py (canonical)
- tensor_product_grid.py → grids/tensor_grid.py (canonical)

This completes Issue #320. The geometry module now has a clean, organized
subdirectory structure with no backward compatibility wrappers.
