+++ mfg_pde/alg/reinforcement/multi_population/base_environment.py
@@ -21,8 +21,8 @@ if TYPE_CHECKING:
 
     from numpy.typing import NDArray
 
+    from mfg_pde.geometry.grids.grid_1d import SimpleGrid1D
     from mfg_pde.geometry.mesh_2d import Mesh2D
-    from mfg_pde.geometry.simple_grid_1d import SimpleGrid1D
 
     from .population_config import PopulationConfig
