+++ mfg_pde/geometry/__init__.py
@@ -99,6 +99,11 @@ from .graph import (
     maze_VoronoiGenerator,
 )
 
+# Grid geometry - Import from subdirectories (canonical locations)
+from .grids.grid_1d import SimpleGrid1D
+from .grids.grid_2d import SimpleGrid2D, SimpleGrid3D
+from .grids.tensor_grid import TensorProductGrid
+
 # Implicit geometry
 from .implicit import (
     ComplementDomain,
@@ -136,12 +141,6 @@ from .protocol import (
     is_geometry_compatible,
     validate_geometry,
 )
-from .simple_grid import SimpleGrid2D, SimpleGrid3D
-
-# Grid geometry
-# Import from old file locations for backward compatibility (Phase 1)
-from .simple_grid_1d import SimpleGrid1D
-from .tensor_product_grid import TensorProductGrid
 
 # Legacy grid imports (from old file names)
