+++ mfg_pde/geometry/amr_1d.py
@@ -28,7 +28,7 @@ if TYPE_CHECKING:
     from mfg_pde.backends.base_backend import BaseBackend
     from mfg_pde.types.solver_types import JAXArray
 
-    from .simple_grid_1d import SimpleGrid1D
+    from .grids.grid_1d import SimpleGrid1D
 
 # Always define JAX_AVAILABLE at module level
 try:
