+++ mfg_pde/utils/sparse_operations.py
@@ -42,7 +42,7 @@ if TYPE_CHECKING:
 
     from numpy.typing import NDArray
 
-    from mfg_pde.geometry.tensor_product_grid import TensorProductGrid
+    from mfg_pde.geometry.grids.tensor_grid import TensorProductGrid
 
 
 class SparseMatrixBuilder:
