+++ tests/unit/geometry/test_tensor_product_grid_ndim.py
@@ -10,7 +10,7 @@ import pytest
 
 import numpy as np
 
-from mfg_pde.geometry.tensor_product_grid import TensorProductGrid
+from mfg_pde.geometry.grids.tensor_grid import TensorProductGrid
 
 
 class TestNDimensionalGrids:
