+++ tests/unit/test_fp_fdm_solver.py
@@ -13,7 +13,7 @@ import numpy as np
 from mfg_pde.alg.numerical.fp_solvers import FPFDMSolver
 from mfg_pde.core.mfg_problem import MFGProblem
 from mfg_pde.geometry import BoundaryConditions
-from mfg_pde.geometry.simple_grid_1d import SimpleGrid1D
+from mfg_pde.geometry.grids.grid_1d import SimpleGrid1D
 
 
 @pytest.fixture
