+++ tests/unit/test_geometry/test_simple_grid.py
@@ -13,7 +13,7 @@ import pytest
 
 import numpy as np
 
-from mfg_pde.geometry.simple_grid import SimpleGrid2D, SimpleGrid3D
+from mfg_pde.geometry.grids.grid_2d import SimpleGrid2D, SimpleGrid3D
 
 # ============================================================================
 # Test: SimpleGrid2D - Initialization
